move quantizeColors to MeshExporter
